Refactor CVE Intelligence Manager and Visual Engine
